backport of commit bcd89f875d88e0a01f2e6f9220f6d72afb5661b2
